requirements.txt
Django==1.11.2
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
node==0.9.22
odict==1.6.2
plumber==1.5
pytz==2017.3
zope.component==4.4.1
zope.deprecation==4.3.0
zope.event==4.3.0
zope.interface==4.4.3
zope.lifecycleevent==4.2.0
